Check prefix folder is correct before deleting it
